+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 ### v0.1.0
+* Restructure controller UI.
+* Add color choosers, for cold, zero and warm temperature
+### v0.1.0
 * Add header line explaining what you see
 * Controller UI starts as Settings icon. When closing controller, the UI is replaced by Settings icon.
 ### v0.0.3